Tools for creating block nodes

These functions are used to create block nodes. Blocks nodes are used to represent block level elements in markdown.

md_block_doc(..., flags = "MD_DIALECT_COMMONMARK")

md_block_quote(...)

md_block_html(...)

md_block_p(...)

md_block_hr(...)

md_block_code(..., info = "", lang = "", fence_char = "`")

md_block_ul(..., tight = 1, mark = "*")

md_block_ol(..., tight = 1L, start = 1, mark_delimiter = ".")

md_block_li(..., is_task = 0, task_mark = " ")

md_block_h(..., level)

md_block_table(..., col_count, body_row_count, head_row_count = 1)

md_block_thead(...)

md_block_tbody(...)

md_block_tr(...)

md_block_th(..., align = c("default", "left", "center", "right"))

md_block_td(..., align = c("default", "left", "center", "right"))

Arguments

...

Child nodes that will be contained within the block - all must inherit from md_node.

flags

Used by md_block_doc to specify the document parser flag(s).

info

Used by md_block_code nodes to specify the code block info string.

lang

Used by md_block_code nodes to specify the code block language.

fence_char

Used by md_block_code nodes to specify the code block fence character.

tight

Used by md_block_ul or md_block_ol nodes to specify whether the list is tight or loose.

mark

Used by md_block_ul nodes to specify the list marker.

start

Used by md_block_ol nodes to specify the start number of the list.

mark_delimiter

Used by md_block_ol nodes to specify the delimiter between the value and text. Only "." and ")" are allowed.

is_task

Used by md_block_li nodes to specify whether the item is a task.

task_mark

Used by md_block_li task nodes to specify the task mark character. Only " ", "x", and "X" are allowed.

level

Used by md_block_h nodes to specify the heading level.

col_count

Used by md_block_table nodes to specify the number of columns.

body_row_count

Used by md_block_table nodes to specify the number of body rows.

head_row_count

Used by md_block_table nodes to specify the number of header rows. Should only be 1.

align

Used by md_block_td or md_block_th nodes to specify the alignment of the table contents.

Value

Returns a list with a class of specified type along with md_span and md_node.

See also