Markdown parser flags

The md4c library supports a number of markdown variants / options. The parsing of these is controlled by flags passed to the parser. The following functions provide commonly used utilities for these flags.

flags_available()

flags_describe()

flags_check(flags, match_case = FALSE)

flags_used(md)

Arguments

flags
Character vector of flag names

match_case
Require flag names to case match

md
Markdown object

Examples

flags_available()

#>  [1] "MD_FLAG_COLLAPSEWHITESPACE"       "MD_FLAG_PERMISSIVEATXHEADERS"    
#>  [3] "MD_FLAG_PERMISSIVEURLAUTOLINKS"   "MD_FLAG_PERMISSIVEEMAILAUTOLINKS"
#>  [5] "MD_FLAG_NOINDENTEDCODEBLOCKS"     "MD_FLAG_NOHTMLBLOCKS"            
#>  [7] "MD_FLAG_NOHTMLSPANS"              "MD_FLAG_TABLES"                  
#>  [9] "MD_FLAG_STRIKETHROUGH"            "MD_FLAG_PERMISSIVEWWWAUTOLINKS"  
#> [11] "MD_FLAG_TASKLISTS"                "MD_FLAG_LATEXMATHSPANS"          
#> [13] "MD_FLAG_WIKILINKS"                "MD_FLAG_UNDERLINE"               
#> [15] "MD_FLAG_PERMISSIVEAUTOLINKS"      "MD_FLAG_NOHTML"                  
#> [17] "MD_DIALECT_COMMONMARK"            "MD_DIALECT_GITHUB"               

flags_describe()

#> # A tibble: 18 x 2
#>    flag                    description                                          
#>    <chr>                   <chr>                                                
#>  1 MD_FLAG_COLLAPSEWHITES… In MD_TEXT_NORMAL, collapse non-trivial whitespace i…
#>  2 MD_FLAG_PERMISSIVEATXH… Do not require space in ATX headers ( ###header )    
#>  3 MD_FLAG_PERMISSIVEURLA… Recognize URLs as autolinks even without '<', '>'    
#>  4 MD_FLAG_PERMISSIVEEMAI… Recognize e-mails as autolinks even without '<', '>'…
#>  5 MD_FLAG_NOINDENTEDCODE… Disable indented code blocks. (Only fenced code work…
#>  6 MD_FLAG_NOHTMLBLOCKS    Disable raw HTML blocks.                             
#>  7 MD_FLAG_NOHTMLSPANS     Disable raw HTML (inline).                           
#>  8 MD_FLAG_TABLES          Enable tables extension.                             
#>  9 MD_FLAG_STRIKETHROUGH   Enable strikethrough extension.                      
#> 10 MD_FLAG_PERMISSIVEWWWA… Enable WWW autolinks (even without any scheme prefix…
#> 11 MD_FLAG_TASKLISTS       Enable task list extension.                          
#> 12 MD_FLAG_LATEXMATHSPANS  Enable $ and $$ containing LaTeX equations.          
#> 13 MD_FLAG_WIKILINKS       Enable wiki links extensn.                           
#> 14 MD_FLAG_UNDERLINE       Enable underline extension (and disables '_' for nor…
#> 15 MD_FLAG_PERMISSIVEAUTO… Combines flags (MD_FLAG_PERMISSIVEEMAILAUTOLINKS | M…
#> 16 MD_FLAG_NOHTML          Combines flags (MD_FLAG_NOHTMLBLOCKS | MD_FLAG_NOHTM…
#> 17 MD_DIALECT_COMMONMARK   All flags off                                        
#> 18 MD_DIALECT_GITHUB       Combines flags (MD_FLAG_PERMISSIVEAUTOLINKS | MD_FLA…